Split GetUserAsync into two methods.
The existing API assumes the user understands that not supplying a login is
equivalent to asking for the current authenticated user which sort of maps
more closely to the URL structure.

I think the api library should hide that implementation detail from the
user and provide explicit methods so it's clear what they should do.