Adds TreeClient.GetRecursive method.
Initially I implemented this as an optional parameter that would append "?recursive=1" to the end of the URI if an optional recursive boolean was supplied to the TreeClient.Get method.  Unfortunately, code analysis complained about that so I went with a set of new overloads instead.

Let me know if a different approach is preferred for adding optional query string parameters to API methods or if this is the preferred solution.  The problem with this solution is that it doesn't scale out well with multiple optional parameters.