Ensure that Commit.Compare(..) sets MergeBaseCommit correctly.
MergeBaseCommit was spelled incorrectly in CompareResult causing Octokit to pass
"null" no matter what the comparison is. This commit makes MergedBaseCommit (the
previous spelling) obsolete, introduces MergeBaseCommit property and adds a test
to ensure that MergeBaseCommit is set properly in CompareResult.

In the test, I specifically compared a commit with itself since the test only
cares whether the property is filled correctly or not, it is not testing whether
the internal computation is correct or not.

Closes #1258.