Use static data for the GetAll tests
The thinking behind this is that to test the pagination works properly, we need a PR with 2 review requests, which takes the number of accounts needed to 3 (the account that creates the PR + the 2 accounts that get a review request).

This would have become too difficult so it was decided to create a test organisation with a repo specifically for this where we'd set up a PR with 2 review requests.