WIP: Add support for querying the Code of Conduct API
While I've contributed to other repositories in the past, this is my first open-source contribution to another C# repository! I *think* this is almost done. I'm working on integration tests.

Design question: Should there be an enum for the Code of Conduct key? Currently there are two officially registered conducts with GitHub, which is the Citizen Conduct and Contributor Covenant.