move deprecation warning from subscribers_count to watchers_count.
Subscribers_count shows the actual watchers_count. Many people think they are retrieving the watchers_count, but are actually retrieving the stargazers_count...
This is because the GitHub UI is still calling subscribers, "Watchers".
To prevent many others from falling into this trap, a warning should be added to the watchers_count property.

On top of this, subscribers_count is no longer deprecated as far as the GitHub documentation goes.

Fixes #2509

Signed-off-by: Brend Smits <brend.smits@philips.com>