fix: Reduce string allocations during SimpleJson.ParseString (#2977)
Noticed this allocation while looking at a profile of solution load in visual studio. These StringBuilder allocations were showing up as 0.5% of total allocations. I took a peek at the code, and reaslized the StringBuilder need not be created unless there is a '\' in the string value being parsed. In that case, just copy already seen characters into a StringBuilder and continue as previously.

Co-authored-by: Nick Floyd <139819+nickfloyd@users.noreply.github.com>